refactor: extract LyricsParser types to types.ts (Sprint 051)
Split 903 LOC LyricsParser: types extracted to types.ts,
index.ts re-exports for backward compatibility.
