+++ src/lib/lyrics/index.ts
@@ -1,24 +1,6 @@
 /**
- * Lyrics utilities export
+ * Lyrics Module
+ * Sprint 051 decomposition: types extracted from LyricsParser.ts
  */
-
-export { LyricsFormatter } from "./LyricsFormatter";
-export type { EnrichmentTags } from "./LyricsFormatter";
-
-export { LyricsValidator } from "./LyricsValidator";
-export type { ValidationResult } from "./LyricsValidator";
-
-export { LyricsParser } from "./LyricsParser";
-export type { ParsedLyrics, LyricsSection, SectionTag } from "./LyricsParser";
-
-// Advanced tag parsing
-export * from "./advancedTagParser";
-
-// Precision synchronization
-export * from "./precisionSync";
-
-// Section matching for replacement
-export * from "./sectionMatcher";
-
-// Unified constants
-export * from "./constants";
+export * from "./types";
+export * from "./LyricsParser";
